+++ guillotina_cms/json/__init__.py
@@ -1,3 +1,4 @@
 from . import serialize_content  # noqa
 from . import serialize_schema  # noqa
 from . import image  # noqa
+from . import db  # noqa