+++ guillotina_cms/subscribers/__init__.py
@@ -1,2 +1,3 @@
 from . import content  # noqa
 from . import edited  # noqa
+from . import db  # noqa