+++ guillotina_cms/subscribers/content.py
@@ -1,17 +1,24 @@
 import datetime
 
 from guillotina import configure
+from guillotina.catalog import index
 from guillotina.component import query_adapter
 from guillotina.interfaces import IObjectAddedEvent
 from guillotina.interfaces import IResource
 from guillotina.security.utils import apply_sharing
+from guillotina.transactions import get_transaction
 from guillotina.utils import get_authenticated_user_id
 from guillotina.utils import get_current_request
+
 from guillotina_cms.interfaces import ICMSBehavior
 from guillotina_cms.interfaces import IWorkflow
+from guillotina_cms.ordering import get_next_order
+from guillotina_cms.ordering import supports_ordering
 
 
-@configure.subscriber(for_=(IResource, IObjectAddedEvent))
+@configure.subscriber(
+    for_=(IResource, IObjectAddedEvent),
+    priority=1001)  # after indexing
 async def cms_object_added(obj, event):
     cms = query_adapter(obj, ICMSBehavior)
     if cms is not None:
@@ -40,6 +47,14 @@ async def cms_object_added(obj, event):
             }
         )
         cms._p_register()
+        txn = get_transaction()
+        if supports_ordering(txn.storage):
+            pos = await get_next_order()
+            cms.position_in_parent = pos
+            fut = index.get_future()
+            if obj.uuid not in fut.index:
+                fut.index[obj.uuid] = {}
+            fut.index[obj.uuid]['position_in_parent'] = cms.position_in_parent
 
     if hasattr(obj, 'title') and obj.title is None:
         obj.title = obj.id